IIOT Project/demos/pymodbus-demo/requirements.txt
pymodbus~=3.5.4
pyserial~=3.5
rich~=13.6.0